chore: refactoring + fixing ESLint

+++ src/wc3maptranslator/HexBuffer.ts
@@ -68,7 +68,7 @@ export class HexBuffer {
     const buf = Buffer.alloc(4)
 
     // ieee754.write(buffer, value, buffer offset, little-endian, mantissa length, bytes);
-    ieee754.write(buf, float, 0, true, 23, 4)
+    ieee754.write(buf as Uint8Array, float, 0, true, 23, 4)
 
     buf.forEach((byte) => {
       this._buffer.push('0x' + byte.toString(16))
